+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.3 / 2022-05-25
+* Updated lib `fs-extra` to `v10.1.0`.
+* Updated lib `oro-functions` to `v1.1.4`.
+* Fixed `this.disconnect()` it's called always with `await`.
+* Fixed `this.mkdir()` behavior because the _RFC 959 standard_ changed.
+* Updated `this.exists` adding second param `disconnectWhenError` to choose it if it's desired.
+
 ## 1.0.2 / 2021-12-14
 * Fixed by default `ftpConfig.readyTimeout` is `3000` miliseconds.